[#1512] Improve ModelMigration implementation
- Update migration to use ModelMigration base class - Refactor migration
code structure for better maintainability - Improve error message for
anonymous customer save attempts - Split migration implementation into
separate files - Remove redundant code and improve readability
