+++ apps/api/v2/models/customer.rb
@@ -393,7 +393,7 @@ module V2
     # This method overrides the default save behavior to prevent
     # anonymous customers from being persisted to the database.
     def save **kwargs
-      raise Onetime::Problem, "Anonymous cannot be saved #{self.class} #{rediskey}" if anonymous?
+      raise Onetime::Problem, "Cannot save anonymous #{self.class} #{identifier}" if anonymous?
 
       super
     end
